+++ lib/exceptions.dart
@@ -1,0 +1,1 @@
+export 'src/command_factory/command_not_defined.exception.dart';