+++ lib/src/command/parameterized_reactive_command/parameterized.reactive.command.dart
@@ -1,0 +1,8 @@
+import '../parameterized_command/parameterized.command.dart';
+import '../reactive_command/reactive.command.dart';
+
+abstract class ParameterizedReactiveCommand<REQUEST extends Object?,
+        RESPONSE extends Object?>
+    implements
+        ParameterizedCommand<REQUEST, Stream<RESPONSE>>,
+        ReactiveCommand<RESPONSE> {}