+++ lib/src/command/reactive_command/reactive.command.dart
@@ -1,0 +1,7 @@
+import '../command.dart';
+
+///
+/// A command that responds with a stream.
+///
+abstract class ReactiveCommand<RESPONSE extends Object?>
+    implements Command<Stream<RESPONSE>> {}