+++ lib/zam_command_pattern.dart
@@ -1,0 +1,4 @@
+export 'src/command/_.index.dart';
+export 'src/command_builder/_.index.dart';
+export 'src/command_executor/_.index.dart';
+export 'src/command_factory/_.index.dart';