+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.1
 
-  * Includes @component_1@.
-  * Includes @component_2@.
-  * Includes @component_3@.
+  * Includes `Command`, `CommandFactory`, `CommandExecutor`, `CommandBuilder`
+  * Includes `AsyncCommand`, `ReactiveCommand`.
+  * Includes `ParameterizedCommand`, `ParameterizedAsyncCommand`, `ParameterizedReactiveCommand`.
+  * Includes `CommandNotDefinedException`.